requirements.txt
openai>=1.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: for --upload-to-drive (Google Drive upload)
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
